Recorded boards creation
Changelog
1. headcount field set to nullable, and no default.
In the update process for single checks, if the most frequent headcount
is None, which means most validations do not include headcount info, we use
the number of candidates when the board was inserted.

If the headcount is 0, which means most frequent headcount think the
candidates are out of our investigative range, we do not about original
relation between boards and candidates.

If headcount > 0, then we update the board-candidates relation, just
like previous commits.

2. Record the inserted boards in boards_checks
When inserting new boards, the board data will be recorded in 'check'
table as well to avoid losing the original description about this board.
This table acts as a history record table.

New field 'is_original' was introduced to distinguish whether this
record is for brand new insertion or a validation.

Migrations are made to record the original data for existing boards.