Fix unfiltered board issue for /board/verify
route: `/api/verify/board`

Changelog
Modified the action when reaching the end of boards records.

Previously when users try to skip the last board in the db,
server will use the default queryset without board id and uploaded_by
restriction. The results will be decided totally on sorting order,
which is `verified_amount ASC` and `uploaded_at ASC`. This will cause
a problem because the last board usually has the least `verified_amount`.
So the second result will still be the last board, which is what we want
to skip.

In this commits, before we re-issue the SQL query, we get the url query
parameters, with which we use to restict the SQL results.